Updates - renaming (#1925)
