+++ source/databricks/calculation_engine/tests/calculation_logic/conftest.py
@@ -14,7 +14,7 @@
 import pytest
 from pyspark.sql import SparkSession
 
-from business_logic_tests.scenario_factory import ScenarioFixture
+from calculation_logic.scenario_factory import ScenarioFixture
 
 
 @pytest.fixture(scope="session")
