+++ source/databricks/calculation_engine/tests/calculation_logic/features/energy_calculations/flex_consumption_per_ga_and_brp_and_es/readme.md
@@ -1,0 +1,15 @@
+# Calculate flex consumption per ga and brp and es
+
+The purpose of this test is to verify that the flex consumption is calculated correctly for
+grid area 804.
+
+Period:
+
+            Period_Start                         Period_End
+             2023-01-31                          2023-02-01
+              23:00:00                            23:00:00
+              ---|-----------------------------------|---
+        
+    MPP1 (E17)   |----------------------------------------------
+    MPP2 (E18)   |----------------------------------------------
+    MPP3 (E17) -------------------------------------------------
