+++ source/databricks/calculation_engine/tests/calculation_logic/features/energy_calculations/flex_consumption_per_ga_and_es/test_calculation_execute.py
@@ -11,16 +11,27 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from pyspark.sql import DataFrame, SparkSession
-
-from business_logic_tests.features.energy_calculations.results_dataframe import (
-    create_result_dataframe,
+from helpers.data_frame_utils import (
+    assert_dataframe_and_schema,
+)
+from .states.scenario_state import (
+    get_expected,
 )
 
 
-def get_expected(*args) -> DataFrame:  # type: ignore
-    spark: SparkSession = args[0]
-    expected_dataframe: DataFrame = args[1]
+def test_execute__returns_expected(  # type: ignore
+    scenario_fixture,
+) -> None:
+    # Arrange
+    scenario_fixture.setup(get_expected)
 
-    return create_result_dataframe(spark, expected_dataframe)
+    # Act
+    results = scenario_fixture.execute()
+
+    # Assert
+    assert_dataframe_and_schema(
+        results.energy_results.flex_consumption_per_ga_and_es.df,
+        scenario_fixture.expected,
+        ignore_decimal_precision=True,
+        ignore_nullability=True,
+    )
