+++ source/databricks/calculation_engine/tests/calculation_logic/features/energy_calculations/flex_consumption_per_ga_and_es/test_data/grid_loss_metering_points.csv
