+++ source/databricks/calculation_engine/tests/calculation_logic/scenario_factory.py
@@ -19,8 +19,8 @@ from unittest.mock import Mock
 
 from pyspark.sql import SparkSession, DataFrame
 
-from business_logic_tests.correlations import get_correlations
-from business_logic_tests.create_calculation_args import create_calculation_args
+from calculation_logic.correlations import get_correlations
+from calculation_logic.create_calculation_args import create_calculation_args
 from package.calculation import PreparedDataReader
 from package.calculation.CalculationResults import (
     CalculationResultsContainer,
